+++ src/network/logout.test.ts
@@ -11,16 +11,19 @@ import * as utils from './utils';
 import server from '../mocks/server';
 import { useLoginConfigStore } from '../store/login/store';
 import { controlConsoleError } from '../tests/utils';
-import type { ErrorSoapResponse } from '../types/network';
 
-const mockEndSession = (response): jest.SpyInstance<HttpResponse> =>
-	jest.spyOn(api, 'endSession').mockResolvedValueOnce(response);
+beforeEach(() => {
+	jest.spyOn(api, 'endSession').mockReturnValueOnce(
+		Promise.resolve({
+			Header: { context: {} },
+			Body: {}
+		})
+	);
+});
 
 describe('Logout', () => {
 	it('should redirect to login page if EndSession request fails', async () => {
 		const goToLoginFn = jest.spyOn(utils, 'goToLogin').mockImplementation();
-		const resp = HttpResponse.json({ error: 'Not Authorized' }, { status: 500 });
-		mockEndSession(resp);
 		await logout();
 		await jest.advanceTimersToNextTimerAsync();
 		expect(goToLoginFn).toHaveBeenCalled();
@@ -28,7 +31,6 @@ describe('Logout', () => {
 
 	it('should redirect to login page if /logout request fails', async () => {
 		const goToLoginFn = jest.spyOn(utils, 'goToLogin').mockImplementation();
-		mockEndSession(HttpResponse.json({}, { status: 500 }));
 		await logout();
 		await jest.advanceTimersToNextTimerAsync();
 		expect(goToLoginFn).toHaveBeenCalled();
@@ -38,7 +40,6 @@ describe('Logout', () => {
 		controlConsoleError('Failed to fetch');
 		const goToLoginFn = jest.spyOn(utils, 'goToLogin').mockImplementation();
 		server.use(http.post('/service/soap/EndSessionRequest', () => HttpResponse.error()));
-		mockEndSession(HttpResponse.error());
 		await logout();
 		await jest.advanceTimersToNextTimerAsync();
 		expect(goToLoginFn).toHaveBeenCalled();
@@ -47,7 +48,6 @@ describe('Logout', () => {
 	it('should redirect to login page if /logout throws error', async () => {
 		controlConsoleError('Failed to fetch');
 		const goToLoginFn = jest.spyOn(utils, 'goToLogin').mockImplementation();
-		mockEndSession(HttpResponse.error());
 		await logout();
 		await jest.advanceTimersToNextTimerAsync();
 		expect(goToLoginFn).toHaveBeenCalled();
@@ -55,28 +55,6 @@ describe('Logout', () => {
 
 	it('should redirect to login page if EndSession request succeeded with Fault', async () => {
 		const goToLoginFn = jest.spyOn(utils, 'goToLogin').mockImplementation();
-		mockEndSession(
-			HttpResponse.json<ErrorSoapResponse>(
-				{
-					Header: { context: {} },
-					Body: {
-						Fault: {
-							Code: { Value: '' },
-							Detail: {
-								Error: {
-									Code: '',
-									Trace: ''
-								}
-							},
-							Reason: {
-								Text: ''
-							}
-						}
-					}
-				},
-				{ status: 200 }
-			)
-		);
 		await logout();
 		await jest.advanceTimersToNextTimerAsync();
 		expect(goToLoginFn).toHaveBeenCalled();
@@ -86,7 +64,6 @@ describe('Logout', () => {
 		it('should redirect to login page if EndSession request fails', async () => {
 			useLoginConfigStore.setState({ carbonioWebUiLogoutURL: 'custom logout' });
 			const goToFn = jest.spyOn(utils, 'goTo').mockImplementation();
-			mockEndSession(HttpResponse.json({}, { status: 500 }));
 			await logout();
 			await jest.advanceTimersToNextTimerAsync();
 			expect(goToFn).toHaveBeenCalled();
@@ -95,7 +72,6 @@ describe('Logout', () => {
 		it('should redirect to login page if /logout request fails', async () => {
 			useLoginConfigStore.setState({ carbonioWebUiLogoutURL: 'custom logout' });
 			const goToFn = jest.spyOn(utils, 'goTo').mockImplementation();
-			mockEndSession(HttpResponse.json({}, { status: 500 }));
 			await logout();
 			await jest.advanceTimersToNextTimerAsync();
 			expect(goToFn).toHaveBeenCalled();
@@ -105,7 +81,6 @@ describe('Logout', () => {
 			useLoginConfigStore.setState({ carbonioWebUiLogoutURL: 'custom logout' });
 			controlConsoleError('Failed to fetch');
 			const goToFn = jest.spyOn(utils, 'goTo').mockImplementation();
-			mockEndSession(HttpResponse.error());
 			await logout();
 			await jest.advanceTimersToNextTimerAsync();
 			expect(goToFn).toHaveBeenCalled();
@@ -115,7 +90,6 @@ describe('Logout', () => {
 			useLoginConfigStore.setState({ carbonioWebUiLogoutURL: 'custom logout' });
 			controlConsoleError('Failed to fetch');
 			const goToFn = jest.spyOn(utils, 'goTo').mockImplementation();
-			mockEndSession(HttpResponse.error());
 			await logout();
 			await jest.advanceTimersToNextTimerAsync();
 			expect(goToFn).toHaveBeenCalled();
@@ -124,28 +98,6 @@ describe('Logout', () => {
 		it('should redirect to login page if EndSession request succeeded with Fault', async () => {
 			useLoginConfigStore.setState({ carbonioWebUiLogoutURL: 'custom logout' });
 			const goToFn = jest.spyOn(utils, 'goTo').mockImplementation();
-			mockEndSession(
-				HttpResponse.json<ErrorSoapResponse>(
-					{
-						Header: { context: {} },
-						Body: {
-							Fault: {
-								Code: { Value: '' },
-								Detail: {
-									Error: {
-										Code: '',
-										Trace: ''
-									}
-								},
-								Reason: {
-									Text: ''
-								}
-							}
-						}
-					},
-					{ status: 200 }
-				)
-			);
 			await logout();
 			await jest.advanceTimersToNextTimerAsync();
 			expect(goToFn).toHaveBeenCalled();