+++ src/settings/components/general-settings/scaling-setting-section.tsx
@@ -7,6 +7,7 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import type { SliderProps } from '@zextras/carbonio-design-system';
 import {
+	FormSubSection,
 	Button,
 	Checkbox,
 	Container,
@@ -24,7 +25,7 @@ import type { ResetComponentImperativeHandler } from '../utils';
 import { getAutoScalingFontSize } from '../utils';
 
 const ScalingSliderContainer = styled(Container)`
-	box-shadow: 0px 0px 4px rgba(166, 166, 166, 0.5);
+	box-shadow: 0 0 4px rgba(166, 166, 166, 0.5);
 	border-radius: 0.5rem;
 `;
 
@@ -191,16 +192,8 @@ export const ScalingSettingSection = ({
 	);
 
 	return (
-		<Container
-			orientation={'vertical'}
-			mainAlignment={'flex-start'}
-			crossAlignment={'flex-start'}
-			gap={'1rem'}
-			height={'fit'}
-			width={'fill'}
-		>
+		<FormSubSection>
 			<Container
-				orientation={'vertical'}
 				gap={'1rem'}
 				height={'fit'}
 				width={'fit'}
@@ -208,7 +201,7 @@ export const ScalingSettingSection = ({
 				crossAlignment={'flex-start'}
 				maxWidth={'100%'}
 			>
-				<Text size="small" overflow={'break-word'}>
+				<Text overflow={'break-word'}>
 					{t(
 						'settings.appearance.labels.choose_size_description',
 						'Choose Type Size and Styles for Carbonio Environment'
@@ -216,7 +209,6 @@ export const ScalingSettingSection = ({
 				</Text>
 				<ScalingSliderContainer
 					orientation="horizontal"
-					width="fill"
 					height="auto"
 					padding={{ vertical: 'small', horizontal: 'large' }}
 					gap={'0.5rem'}
@@ -272,17 +264,14 @@ export const ScalingSettingSection = ({
 						'settings.appearance.labels.choose_size_auto',
 						'Automatically resize the text size depending on the device'
 					)}
-					size={'small'}
 				/>
 			</Container>
 			<ExampleContainer
 				background={'gray5'}
 				padding={'large'}
 				gap={'0.5rem'}
-				orientation={'vertical'}
 				mainAlignment={'flex-start'}
 				crossAlignment={'flex-start'}
-				width={'fill'}
 				$fontSize={exampleFontSize}
 			>
 				<ExampleText weight={'light'} overflow={'break-word'}>
@@ -295,6 +284,6 @@ export const ScalingSettingSection = ({
 					{exampleText}
 				</ExampleText>
 			</ExampleContainer>
-		</Container>
+		</FormSubSection>
 	);
 };